drive

+++ server.js
@@ -553,7 +553,7 @@ app.get('/files/realizados', function(req, res){
           
         }); //end query
         query.on('end',function(){
-            var query2 = client.query('select * from pedidos where finalizado=$1 and idpedido=$2 order by idpedido limit $3',[true,parseInt(idCorte),limit], function(err, result) {       
+            var query2 = client.query('select * from pedidos where finalizado=$1 and idpedido=$2 order by idpedido limit $3',[true,parseFloat(idCorte),limit], function(err, result) {       
             if (err)
              { console.error(err);}
             else
